Politics Tendences
Here you can find a entire project that I developed a question about: ¿It's only Metropolitan Region* is representative about the political tendencies in other regions? we focused in the behavior of this region and we compared with the others.


Metropolitan Region: Includes Santiago of Chile, largest city in the country.